backend/flask-api/requirements.txt
Flask
tensorflow
Pillow
requests
google-generativeai
python-dotenv
fuzzywuzzy